Generalize Renderer
I stopped liking the idea of it only being able to render a grid. Also, it
now needs the screen width and height for construction. I played with the
idea of it also getting a function that swapped the window buffer but I
felt it complicated things instead of simplifying.